refactor(codegen): split EVM IR passes
